Revert "fix memory access issue on raspbi"
This reverts commit ebb094ad

+++ backend/board/domain/player.go
@@ -3,6 +3,6 @@ package boardDomain
 type Player struct {
 	Id string `json:"id"`
 	// TODO maybe change position to field
-	Position uint8  `json:"position"`
-	Balance  uint32 `json:"balance"`
+	Position int `json:"position"`
+	Balance  int `json:"balance"`
 }
